pyproject.toml
[project]
name = "ai-resume-agent"
version = "0.1.0"
description = "AI-powered agent to tailor resumes based on job descriptions using local LLMs via Ollama"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "pydantic[email]>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
]
